feat: :sparkles: Compartir tareas entre usuarios

+++ resources/views/livewire/task-component.blade.php
@@ -1,5 +1,5 @@
 <!-- ====== Table Section Start -->
-<section class="bg-white ">
+<section class="bg-white " wire:poll="renderAllTasks">
     <div class="container" >
 
        <div class="flex flex-wrap -mx-4">
@@ -77,6 +77,7 @@
                             "
                             >
                             {{ $task->title }}
+
                         </td>
                         <td
                             class="
@@ -102,12 +103,28 @@
                             border-b border-[#E8E8E8]
                             "
                             >
-                            <button wire:click="openCreateModal({{ $task }})" class="bg-yellow-800 text-white px-4 py-2 rounded-md hover:bg-yellow-700">
-                                Editar
-                            </button>
-                            <button wire:click="deleteTask({{ $task }})" wire:confirm="Are you sure you want to delete this task?" class="bg-red-800 text-white px-4 py-2 rounded-md hover:bg-red-700">
-                                Borrar
-                            </button>
+                            @if ((isset($task->pivot) ))
+                                    <button wire:click="taskUnshared({{ $task }})" wire:confirm="Estas seguro que quieres realizar la accion?" class="bg-blue-800 text-white px-4 py-2 rounded-md hover:bg-blue-700">
+                                        DES
+                                    </button>
+
+                            @endif
+                            @if( (isset($task->pivot) && $task->pivot->permission == 'edit') ||(auth()->user()->id == $task->user_id))
+                                <div class="flex flex-row justify-between">
+                                    <button wire:click="openCreateModal({{ $task }})" class="bg-yellow-800 text-white px-4 py-2 rounded-md hover:bg-yellow-700">
+                                        Editar
+                                    </button>
+
+                                    <button wire:click="openShareModal({{ $task }})" class="bg-purple-800 text-white px-2 py-2 rounded-md hover:bg-purple-700">
+                                        Compartir
+                                    </button>
+                                    <button wire:click="deleteTask({{ $task }})" wire:confirm="Are you sure you want to delete this task?"  class="bg-red-800 text-white px-4 py-2 rounded-md hover:bg-red-700">
+                                        Borrar
+                                    </button>
+
+                                </div>
+
+                            @endif
                         </td>
 
                         </tr>
@@ -151,6 +168,50 @@
 
                     </button>
                     <button class="p-3 bg-white border rounded-full w-full font-semibold" wire:click.prevent="closeCreateModal">Cancelar</button>
+                </div>
+                </div>
+            </div>
+            </div>
+        </div>
+
+    @endif
+    @if ($modalShare)
+        <div class="fixed left-0 top-0 flex h-full w-full items-center justify-center bg-black bg-opacity-50 py-10">
+            <div class="max-h-full w-full max-w-xl overflow-y-auto sm:rounded-2xl bg-white">
+            <div class="w-full">
+                <div class="m-8 my-20 max-w-[400px] mx-auto">
+                <div class="mb-8">
+                    <h1 class="mb-4 text-3xl font-extrabold">Compartir tarea</h1>
+
+                    <form>
+                        <div class="mb-4">
+                            <label for="title" class="block mb-2 text-sm font-medium text-gray-900 ">Usuario a compartir</label>
+                            <select wire:model="user_id" class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500">
+                                <option value="">Seleccione un usuario</option>
+                                @foreach ($users as $user)
+                                    <option value="{{ $user->id }}">{{ $user->name }}</option>
+                                @endforeach
+                            </select>
+                        </div>
+                        <div class="mb-4">
+                            <label for="description" class="block mb-2 text-sm font-medium text-gray-900 ">Permisos para compartir</label>
+                            <select wire:model="permiso" class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500">
+                                <option value="">Seleccione un permiso</option>
+                                <option value="view">Ver</option>
+                                <option value="edit">Editar</option>
+                            </select>
+                        </div>
+                    </form>
+
+                </div>
+
+                <div class="flex flex-row">
+                    <button class="p-3 bg-black rounded-full text-white w-full font-semibold"
+                    wire:click="shareTask"
+                >Compartir tarea</button>
+
+                    </button>
+                    <button class="p-3 bg-white border rounded-full w-full font-semibold" wire:click.prevent="closeShareModal">Cancelar</button>
                 </div>
                 </div>
             </div>
